Allow pandas >2 dependency

+++ causallib/tests/test_causal_simulator3.py
@@ -350,7 +350,7 @@ class CS3TestCase(unittest.TestCase):
                   outcome_types=self.no_X.outcome_types, snr=snr, effect_sizes=self.no_X.effect_sizes)
         X, prop, cf = sim.generate_data(num_samples=self.NUM_SAMPLES)
 
-        singular_values = np.linalg.svd(X.values, compute_uv=False)
+        singular_values = np.linalg.svd(X.astype(float).values, compute_uv=False)
         eps = 1e-10
         rank = np.sum(singular_values > eps)
         self.assertEqual(rank, 2,
